refactor: order의 orderTable 검증 로직 분리
