+++ src/main/java/kitchenpos/order/domain/OrderValidator.java
@@ -1,35 +1,25 @@
 package kitchenpos.order.domain;
 
 import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.domain.OrderTableRepository;
+import kitchenpos.order.infrastructure.OrderTableValidator;
 import org.springframework.stereotype.Component;
 
 @Component
 public class OrderValidator {
 
     private final MenuRepository menuRepository;
-    private final OrderTableRepository orderTableRepository;
+    private final OrderTableValidator orderTableValidator;
 
-    public OrderValidator(MenuRepository menuRepository, OrderTableRepository orderTableRepository) {
+    public OrderValidator(MenuRepository menuRepository, OrderTableValidator orderTableValidator) {
         this.menuRepository = menuRepository;
-        this.orderTableRepository = orderTableRepository;
+        this.orderTableValidator = orderTableValidator;
     }
 
     public void validate(Long orderTableId, OrderLineItems orderLineItems) {
-        OrderTable orderTable = getOrderTable(orderTableId);
-
-        if (orderTable.isEmpty()) {
-            throw new IllegalArgumentException("빈 테이블입니다.");
-        }
+        orderTableValidator.validateEmptyTable(orderTableId);
 
         if (orderLineItems.size() != menuRepository.countByIdIn(orderLineItems.getMenuIds())) {
             throw new IllegalArgumentException("주문항목의 수와 메뉴의 수가 일치하지 않습니다.");
         }
-    }
-
-    private OrderTable getOrderTable(Long orderTableId) {
-        return orderTableRepository.findById(orderTableId)
-                .orElseThrow(() -> new IllegalArgumentException("주문 테이블이 존재하지 않습니다."));
     }
 }
