+++ src/main/java/kitchenpos/order/infrastructure/OrderTableValidator.java
@@ -1,0 +1,9 @@
+package kitchenpos.order.infrastructure;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public interface OrderTableValidator {
+
+    void validateEmptyTable(Long orderTableId);
+}
